use string replace loader to add component import
