+++ src/browser.js
@@ -5,6 +5,7 @@ import { render } from 'preact';
 import { hydrate as hydrateCSS } from 'emotion';
 import createStore from 'unistore';
 import { Provider } from 'unistore/preact';
+import Page from '__PAGE__ENTRY__POINT__';
 
 // webpack-specific
 // eslint-disable-next-line camelcase,no-undef
@@ -12,13 +13,16 @@ __webpack_public_path__ = '/assets/javascript/';
 
 const { data, cssIDs } = window.gu.app;
 
-export default Page => {
-    hydrateCSS(cssIDs);
-    render(
-        <Provider store={createStore(data)}>
-            <Page />
-        </Provider>,
-        document.getElementById('app'),
-        document.getElementById('app').lastElementChild,
-    );
-};
+if (module.hot) {
+    module.hot.accept();
+    require('preact/debug'); // eslint-disable-line global-require
+}
+
+hydrateCSS(cssIDs);
+render(
+    <Provider store={createStore(data)}>
+        <Page />
+    </Provider>,
+    document.getElementById('app'),
+    document.getElementById('app').lastElementChild,
+);
