--- src/components/pages/Article.immersive.js
@@ -1,5 +1,0 @@
-// @flow
-
-import Header from 'components/Header';
-
-export default () => <Header />;
